requirements.txt
pandas==2.0.3
biopython==1.81
numpy==1.24.3